pyproject.toml
[project]
name = "sensei"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coverage>=7.12.0",
    "fastapi>=0.123.7",
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.23.1",
    "packaging>=25.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sqlite-vec>=0.1.6",
    "sse-starlette>=3.0.3",
    "typer>=0.20.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "nuitka>=2.8.9",
    "patchelf>=0.17.2.4",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
]
